requirements.txt
streamlit>=1.36
openai>=1.44.0
python-dotenv>=1.0.1
feedparser>=6.0.11
trafilatura>=1.9.0
beautifulsoup4>=4.12.3
requests>=2.32.3
APScheduler>=3.10.4
pydantic>=2.9.2
yfinance>=0.2.44
pandas>=2.2.2
numpy>=2.1.1
SQLAlchemy>=2.0.35
vaderSentiment>=3.3.2
pytz>=2024.1
requests>=2.32.3
